# swagger/api.yaml
#
# For a free graphical editor, try https://mermade.github.io/openapi-gui/#
#
openapi: 3.0.0
servers:
  - url: 'http://jargonbusterbackend.azurewebsites.net/api'
    description: ''
info:
  description: >-
    API for the JargonBuster challenge. Provides a set of APIs to use text
    summarization, medical text analytics and access to other cognitive
    services. 
  version: 1.0.0
  title: JargonBuster API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    name: Jargonbuster Backend
    url: 'https://github.com/faxg/jargonbuster-backend'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
tags:
  - name: Extract
    description: 'API to extract text from documents (like PDF, images)'
  - name: Summarize
    description: Summarize text using different auto-summarization algorithms
  - name: Analyze
    description: >-
      Analytics on medical texts. extracts concepts, entities, key phrases
      etc...
  - name: Present
    description: >-
      Presentation related functions, e.g. find illustrations and, configure
      Azure Immersive Reader 
paths:
  /summarize:
    post:
      summary: Summarizes a text
      description: Summarize text up to 1MB.
      parameters:
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: Specifies the language of the text (default is english)
      operationId: app.post_summary
      responses:
        '200':
          description: 'Summarization successful '
          content:
            application/json:
              schema: {}
      tags:
        - Summarize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummaryRequest'
        description: SummaryRequest containing the text to summarize.
  /extract:
    summary: Extracts text from documents
    description: 'Upload a document (PDF, for now) and extract the text from it.'
    post:
      tags:
        - Extract
      operationId: app.post_extract
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        description: Extract text from a document (PDF for now)
      responses:
        '200':
          description: File uploaded.
        '400':
          description: Error during file upload.
      parameters: []
  /analyze:
    post:
      summary: Analyzes medical texts
      description: Response with detailed analytics of the uploaded text
      parameters: []
      operationId: app.post_analyze
      responses:
        '200':
          description: Text extraction successfull
          content:
            application/json:
              schema: {}
      tags:
        - Analyze
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
        description: 'Text to analyze. '
    summary: Analyze medical jargon text
    description: >-
      This uses Azure Text Analytics for Health to find concepts, terms etc. in
      medical literature
  /getIRToken:
    get:
      summary: Gets a reader token for Azure Immersive Reader
      description: >-
        Clients need to request a valid token for embedding the Immersive Reader
        component. Use this to request one.
      parameters: []
      operationId: app.get_ir_token
      responses:
        '200':
          description: IR reader token acquired
          content:
            application/json:
              schema: {}
      tags:
        - Present
    summary: Presentation helpers
    description: >-
      Helper methods to support the frontend with presenting content in a web /
      mobile user interface.
  /definition:
    get:
      summary: Lookup the definition of a medical term 
      description: >-
        Looks up a medical term definition in the Merriam-Webster medical dictionary
      parameters:
        - name: term
          in: query
          required: false
          schema:
            type: string
          description: the (medical) term to look up
      operationId: app.get_definition
      responses:
        '200':
          description: request ok
          content:
            application/json:
              schema: {}
      tags:
        - Present
    summary: Presentation helpers
    description: >-
      Helper methods to support the frontend with presenting content in a web /
      mobile user interface.
components:
  schemas:
    SummaryRequest:
      type: object
      properties:
        text:
          type: string
    AnalysisRequest:
      type: object
      properties:
        text:
          type: string
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header
  links: {}
  callbacks: {}
security: []